LayoutTests/platform/glib/imported/w3c/web-platform-tests/xhr/access-control-basic-get-fail-non-simple-expected.txt

CONSOLE MESSAGE: Preflight response is not successful. Status code: 400
CONSOLE MESSAGE: XMLHttpRequest cannot load http://www1.web-platform.test:8800/xhr/resources/access-control-basic-options-not-supported.py due to access control checks.

PASS Preflighted cross-origin request denied
